refactor: 잔돈 충전 탭 기능 관련 코드를 mvc 디자인 패턴으로 수정
