--- src/js/models/getRandomCoin.js
@@ -1,5 +1,0 @@
-const getRandomCoin = () => {
-  return MissionUtils.Random.pickNumberInList([500, 100, 50, 10]);
-};
-
-export default getRandomCoin;
