+++ src/js/views/resetInput.js
@@ -1,9 +1,5 @@
 import $ from '../utils/dom.js';
 
-export const resetChargeInput = () => {
-  $('#vending-machine-charge-input').value = '';
-};
-
 export const resetPurchaseInput = () => {
   $('#charge-input').value = '';
 };
