+++ src/js/views/vendingMachineView.js
@@ -1,7 +1,11 @@
 import $ from '../utils/dom.js';
 import store from '../utils/store.js';
 
-const printChargeMoney = () => {
+export const resetChargeInput = () => {
+  $('#vending-machine-charge-input').value = '';
+};
+
+export const printChargeMoney = () => {
   const coins = store.getLocalStorage('coins');
   $('#vending-machine-charge-amount').innerText = `${coins.amount}원`;
   $('#vending-machine-coin-500-quantity').innerText = `${coins.fiveHundred}개`;
@@ -9,5 +13,3 @@ const printChargeMoney = () => {
   $('#vending-machine-coin-50-quantity').innerText = `${coins.fifty}개`;
   $('#vending-machine-coin-10-quantity').innerText = `${coins.ten}개`;
 };
-
-export default printChargeMoney;
